initial commit on CMake branch
Added a .gitignore file, and moved module sources into a nested melatonin_perfetto directory. This is because JUCE requires that the directory name is always the same as the module name, and some tools may change the name of the directory they clone the source code into (CMake's FetchContent and CPM.cmake both do this, they will clone the code into a directory named something like melatonin_perfetto-src, which will then mess up JUCE's CMake code when adding the module with the name melatonin_perfetto).